Put the pydoctor customization into it's own module
